separate model / persistence specific code from rest
