+++ rasa/architecture_prototype/graph.py
@@ -1,29 +1,20 @@
-import abc
 import copy
-import json
-import os
-import tarfile
-from abc import ABC
 from collections import ChainMap
 import inspect
-from pathlib import Path
 from typing import Any, Text, Dict, List, Union, TYPE_CHECKING, Tuple, Optional
 
 import dask
 
 from rasa.architecture_prototype.graph_fingerprinting import TrainingCache
-from rasa.core.channels import UserMessage
+from rasa.architecture_prototype.persistence import AbstractModelPersistor
 from rasa.shared.constants import DEFAULT_DATA_PATH
 import rasa.shared.utils.common
 import rasa.utils.common
 import rasa.core.training
-from rasa.shared.core.domain import Domain
-from rasa.shared.core.events import UserUttered
-from rasa.shared.core.trackers import DialogueStateTracker
 import rasa.shared.utils.io
 
 if TYPE_CHECKING:
-    from rasa.core.policies.policy import PolicyPrediction
+    pass
 
 
 class RasaComponent:
@@ -137,158 +128,7 @@ class RasaComponent:
         return f"{self._component_class}.{self._fn_name}"
 
 
-class ComponentPersistor:
-    def __init__(self, node_name: Text, parent_dir: Path) -> None:
-        self._node_name = node_name
-        self._parent_dir = parent_dir
-        self._dir_for_node = Path(parent_dir / node_name)
-
-    def file_for(self, filename: Text) -> Text:
-        self._dir_for_node.mkdir(exist_ok=True)
-        return str(self._dir_for_node / filename,)
-
-    def directory_for(self, dir_name: Text) -> Text:
-        self._dir_for_node.mkdir(exist_ok=True)
-        directory = self._dir_for_node / dir_name
-        directory.mkdir()
-        return str(directory)
-
-    def get_resource(self, resource_name, filename) -> Text:
-        return str(Path(self._parent_dir, resource_name, filename))
-
-    def resource_name(self) -> Text:
-        return self._node_name
-
-
-class AbstractModelPersistor(ABC):
-    @abc.abstractmethod
-    def create_component_persistor(self, node_name: Text) -> "ComponentPersistor":
-        raise NotImplementedError("Please implement this.")
-
-    @abc.abstractmethod
-    def persist_model(
-        self, target: Text, predict_graph_schema: Dict[Text, Any],
-    ) -> None:
-        raise NotImplementedError("Please implement this.")
-
-    @abc.abstractmethod
-    def load_model(self, persisted_model: Text) -> "Model":
-        raise NotImplementedError("Please implement this.")
-
-    @classmethod
-    def default(cls) -> "AbstractModelPersistor":
-        return LocalModelPersistor(Path("model"))
-
-
-class LocalModelPersistor(AbstractModelPersistor):
-    def __init__(self, local_path: Path = Path("model")) -> None:
-        self._dir = local_path
-
-    def create_component_persistor(self, node_name: Text) -> "ComponentPersistor":
-        return ComponentPersistor(node_name, self._dir)
-
-    def persist_model(
-        self, target: Text, predict_graph_schema: Dict[Text, Any],
-    ) -> None:
-        graph = serialize_graph_schema(predict_graph_schema)
-        (self._dir / "predict_graph.json").write_text(graph)
-
-        with tarfile.open(target, "w:gz") as tar:
-            for elem in os.scandir(self._dir):
-                tar.add(elem.path, arcname=elem.name)
-
-    def load_model(self, persisted_model: Text) -> "Model":
-        with tarfile.open(persisted_model, mode="r:gz") as tar:
-            tar.extractall(self._dir)
-
-        graph = deserialize_graph_schema((self._dir / "predict_graph.json").read_text())
-
-        return Model(graph, self)
-
-
-class ModelTrainer:
-    def __init__(
-        self, model_persistor: AbstractModelPersistor, cache: TrainingCache,
-    ) -> None:
-        self._model_persistor = model_persistor
-        self._cache = cache
-
-    def train(
-        self,
-        # TODO: add "project"
-        train_graph_schema: Dict[Text, Any],
-        predict_graph_schema: Dict[Text, Any],
-    ) -> "Model":
-        fill_defaults(train_graph_schema)
-        core_targets = [
-            "train_memoization_policy",
-            "train_ted_policy",
-            "train_rule_policy",
-        ]
-        nlu_targets = [
-            "train_classifier",
-            "train_response_selector",
-            "train_synonym_mapper",
-        ]
-
-        run_as_dask_graph(
-            train_graph_schema,
-            core_targets + nlu_targets,
-            cache=self._cache,
-            model_persistor=self._model_persistor,
-        )
-
-        fill_defaults(predict_graph_schema)
-        return Model(predict_graph_schema, persistor_factory=self._model_persistor)
-
-
-class Model:
-    def __init__(
-        self, rasa_graph: Dict[Text, Any], persistor_factory: "AbstractModelPersistor"
-    ) -> None:
-        self._rasa_graph = rasa_graph
-        self._predict_graph = convert_to_dask_graph(
-            self._rasa_graph, model_persistor=persistor_factory
-        )
-
-    def handle_message(
-        self, tracker: DialogueStateTracker, message: Optional[UserMessage]
-    ) -> Tuple["PolicyPrediction", Optional[UserUttered]]:
-        graph = self._predict_graph.copy()
-
-        # Insert user message into graph
-        graph["load_user_message"] = _graph_component_for_config(
-            "load_user_message",
-            self._rasa_graph["load_user_message"],
-            {"message": message},
-        )
-
-        # Insert dialogue history into graph
-        graph["load_history"] = _graph_component_for_config(
-            "load_history", self._rasa_graph["load_history"], {"tracker": tracker}
-        )
-
-        prediction, tracker_with_user_event = dask.get(
-            graph, ["select_prediction", "add_parsed_nlu_message"]
-        )
-
-        user_event = None
-        if message:
-            user_event = tracker_with_user_event["add_parsed_nlu_message"].events[-1]
-
-        return prediction["select_prediction"], user_event
-
-    def predict_next_action(
-        self, tracker: DialogueStateTracker,
-    ) -> Tuple["PolicyPrediction", UserUttered]:
-        return self.handle_message(tracker, message=None)
-
-    def get_domain(self) -> Domain:
-        domain_graph = _minimal_graph(self._predict_graph, targets=["load_domain"])
-        return dask.get(domain_graph, "load_domain")["load_domain"]
-
-
-def _minimal_graph(
+def minimal_dask_graph(
     dask_graph: Dict[Text, Tuple[RasaComponent, Text]], targets: List[Text]
 ) -> Dict[Text, Tuple[RasaComponent, Text]]:
     dependencies = _all_dependencies(dask_graph, targets)
@@ -347,13 +187,13 @@ def convert_to_dask_graph(
 
     dsk = {}
     for step_name, step_config in graph_schema.items():
-        dsk[step_name] = _graph_component_for_config(
+        dsk[step_name] = graph_component_for_config(
             step_name, step_config, cache=cache, model_persistor=model_persistor
         )
     return dsk
 
 
-def _graph_component_for_config(
+def graph_component_for_config(
     step_name: Text,
     step_config: Dict[Text, Any],
     config_overrides: Dict[Text, Any] = None,
@@ -384,17 +224,6 @@ def _graph_component_for_config(
     )
 
 
-def fill_defaults(graph_schema: Dict[Text, Any]):
-    for step_name, step_config in graph_schema.items():
-        component_class = step_config["uses"]
-
-        if hasattr(component_class, "defaults"):
-            step_config["config"] = {
-                **component_class.defaults,
-                **step_config["config"],
-            }
-
-
 def run_as_dask_graph(
     graph_schema: Dict[Text, Any],
     target_names: Union[Text, List[Text]],
@@ -417,39 +246,3 @@ def run_dask_graph(
 def visualise_as_dask_graph(graph_schema: Dict[Text, Any], filename: Text) -> None:
     dask_graph = convert_to_dask_graph(graph_schema)
     dask.visualize(dask_graph, filename=filename)
-
-
-def serialize_graph_schema(graph_schema: Dict[Text, Any]) -> Text:
-    to_serialize = copy.copy(graph_schema)
-    for step_name, step_config in to_serialize.items():
-        component_class = step_config["uses"]
-        step_config["uses"] = component_class.name
-    return json.dumps(to_serialize)
-
-
-def deserialize_graph_schema(
-    serialized_graph_schema: Union[Text, Path]
-) -> Dict[Text, Any]:
-    from rasa.architecture_prototype.graph_components import load_graph_component
-    import rasa.nlu.registry
-    import rasa.core.registry
-
-    schema = json.loads(serialized_graph_schema)
-    for step_name, step_config in schema.items():
-        component_class_name = step_config["uses"]
-        try:
-            step_config["uses"] = rasa.nlu.registry.get_component_class(
-                component_class_name
-            )
-        except Exception:
-            try:
-                step_config["uses"] = load_graph_component(component_class_name)
-            except Exception:
-                try:
-                    step_config["uses"] = rasa.core.registry.policy_from_module_path(
-                        component_class_name
-                    )
-                except:
-                    raise ValueError(f"Unknown component: {component_class_name}")
-
-    return schema
